+++ src/index.ts
@@ -1,6 +1,3 @@
-// Abstract Hue Server
-export * from './abstract-hue-server';
-
 // UPNP
 export * from './upnp/hue-upnp';
 
@@ -12,15 +9,14 @@ export * from './server/hue-server-callback';
 export * from './builder/hue-builder';
 export * from './builder/host';
 export * from './builder/port';
+export * from './builder/https';
+export * from './builder/https-config';
 export * from './builder/discovery-host';
 export * from './builder/discovery-port';
 export * from './builder/udn';
 
 // Error
 export * from './error/hue-error';
-
-// Events
-export * from './events/pairing-event';
 
 // Logger
 export * from './logger';